requirements.txt
PyYAML==6.0
requests==2.28.2
schedule==1.1.0
selenium==4.8.2
